Refactor render pipeline helpers
